requirements.txt
Django==1.9.7
MySQL-python<2.0
django-environ==0.4.0
djangorestframework==3.4.0
django-webpack-loader==0.3.3
six==1.10.0
wsgiref==0.1.2
